test: Add ExpectDownloadFile.download_string() to simplify tests
